Add files via upload
The purpose of the project is to investigate the performance of the human brain using PET imaging. This performance will be examined by identifying active voxels. In addition, the brain's performance will also be studied using SVD.

We have been provided with 12 volumetric images of the human brain, each with dimensions of 876526. The task performed is called "finger opposition." Therefore, out of the available 12 images, 6 of them are in a resting state, and the other 6 are in a task-performing state. The data present in the individual images constitute the task state, while the resting state is represented in the paired images.